# Improve ConcurrentDictionary performance, in particular for strings