- By default, string hash codes are randomized.  This is an important defense-in-depth security measure, but it also adds some overhead when strings are used as keys in dictionaries.  `Dictionary<>` addresses that overhead by starting out with using non-randomized comparers, and then upgrades to randomized comparers only once enough collisions have been detected. This PR updates `ConcurrentDictionary<>` with similar tricks.  The comparer is moved from being stored on the dictionary itself to instead be stored on the Tables object that's atomically swapped when the table grows; that way, the comparer always remains in sync with the hashcodes stored in the nodes in that table.  When we enumerate a bucket looking for an existing key as part of an add, we count how many items we traverse, and if that resulting number is larger than the allowed threshold and we're currently using a non-randomized comparer, we force a rehash; that rehash will replace the non-randomized comparer with the equivalent randomized one.
- The `ConcurrentDictionary<>` ctor is improved to presize based on the size of a collection being passed in; otherwise, it might resize multiple times as it's populating the dictionary.  The sizing logic is also changed to use the same prime bucket selection size as does `Dictionary<>`.
- The method we were using to compute the bucket for a key wasn't being inlined for reference type keys due to the generic context; that method has been moved to the outer type as a static to avoid the non-inlined call and extra generic dictionary lookup.
- For all key types, we were also paying for a non-inlined ldelema helper call when reading the head node of a bucket; that's been addressed via a struct wrapper with a volatile node field, rather than using Volatile.Read to access the array element.
- We were inconsistent in whether checked math was used in computing the size of the table.  In some situations it would be possible to overflow without it being detected, or for it to be detected and manifest in various ways.  This simplifies to just always use checked for computing the counts.
- Remove unnecessary try/finally blocks that are leftover from CERs and thread abort protection.
- Deduped some code with calls to helper functions.

|       Type |      Method |Toolchain | CustomComparer | IgnoreCase |       Mean |    Error |   StdDev | Ratio |
|----------- |------------ |--------- |--------------- |----------- |-----------:|---------:|---------:|------:|
|   GuidKeys | TryGetValue | \main    |              ? |          ? |   778.9 ns | 13.28 ns | 12.42 ns |  1.00 |
|   GuidKeys | TryGetValue |   \pr    |              ? |          ? |   565.5 ns |  8.36 ns |  7.82 ns |  0.73 |
|            |             |          |                |            |            |          |          |       |
| ObjectKeys | TryGetValue | \main    |          False |          ? | 1,139.3 ns | 19.87 ns | 17.61 ns |  1.00 |
| ObjectKeys | TryGetValue |   \pr    |          False |          ? | 1,103.1 ns | 11.62 ns | 10.87 ns |  0.97 |
|            |             |          |                |            |            |          |          |       |
| ObjectKeys | TryGetValue | \main    |           True |          ? | 1,122.4 ns |  9.74 ns |  9.11 ns |  1.00 |
| ObjectKeys | TryGetValue |   \pr    |           True |          ? | 1,050.3 ns | 11.08 ns | 10.36 ns |  0.94 |
|            |             |          |                |            |            |          |          |       |
| StringKeys | TryGetValue | \main    |              ? |       True | 2,331.2 ns | 25.81 ns | 22.88 ns |  1.00 |
| StringKeys | TryGetValue |   \pr    |              ? |       True | 1,359.4 ns | 24.38 ns | 22.81 ns |  0.58 |
|            |             |          |                |            |            |          |          |       |
| StringKeys | TryGetValue | \main    |              ? |      False | 2,616.4 ns | 41.60 ns | 36.88 ns |  1.00 |
| StringKeys | TryGetValue |   \pr    |              ? |      False | 1,826.4 ns | 18.59 ns | 15.52 ns |  0.70 |

```C#
using BenchmarkDotNet.Attributes;
using BenchmarkDotNet.Running;
using System;
using System.Collections.Concurrent;
using System.Collections.Generic;
using System.Linq;

[MemoryDiagnoser(false)]
public partial class Program
{
    static void Main(string[] args) => BenchmarkSwitcher.FromAssembly(typeof(Program).Assembly).Run(args);
}

[MemoryDiagnoser(false)]
public class StringKeys
{
    private KeyValuePair<string, string>[] _pairs;
    private ConcurrentDictionary<string, string> _cd;

    [Params(false, true)]
    public bool IgnoreCase { get; set; }

    [GlobalSetup]
    public void Setup()
    {
        _pairs =
            // from https://github.com/dotnet/runtime/blob/a30de6d40f69ef612b514344a5ec83fffd10b957/src/libraries/System.Formats.Asn1/src/System/Formats/Asn1/WellKnownOids.cs#L317-L419
            new[]
            {
                "1.2.840.10040.4.1", "1.2.840.10040.4.3", "1.2.840.10045.2.1", "1.2.840.10045.1.1", "1.2.840.10045.1.2", "1.2.840.10045.3.1.7", "1.2.840.10045.4.1", "1.2.840.10045.4.3.2", "1.2.840.10045.4.3.3", "1.2.840.10045.4.3.4",
                "1.2.840.113549.1.1.1", "1.2.840.113549.1.1.5", "1.2.840.113549.1.1.7", "1.2.840.113549.1.1.8", "1.2.840.113549.1.1.9", "1.2.840.113549.1.1.10", "1.2.840.113549.1.1.11", "1.2.840.113549.1.1.12", "1.2.840.113549.1.1.13",
                "1.2.840.113549.1.5.3", "1.2.840.113549.1.5.10", "1.2.840.113549.1.5.11", "1.2.840.113549.1.5.12", "1.2.840.113549.1.5.13", "1.2.840.113549.1.7.1", "1.2.840.113549.1.7.2", "1.2.840.113549.1.7.3", "1.2.840.113549.1.7.6",
                "1.2.840.113549.1.9.1", "1.2.840.113549.1.9.3", "1.2.840.113549.1.9.4", "1.2.840.113549.1.9.5", "1.2.840.113549.1.9.6", "1.2.840.113549.1.9.7", "1.2.840.113549.1.9.14", "1.2.840.113549.1.9.15", "1.2.840.113549.1.9.16.1.4",
                "1.2.840.113549.1.9.16.2.12", "1.2.840.113549.1.9.16.2.14", "1.2.840.113549.1.9.16.2.47", "1.2.840.113549.1.9.20", "1.2.840.113549.1.9.21", "1.2.840.113549.1.9.22.1", "1.2.840.113549.1.12.1.3", "1.2.840.113549.1.12.1.5",
                "1.2.840.113549.1.12.1.6", "1.2.840.113549.1.12.10.1.1", "1.2.840.113549.1.12.10.1.2", "1.2.840.113549.1.12.10.1.3", "1.2.840.113549.1.12.10.1.5", "1.2.840.113549.1.12.10.1.6", "1.2.840.113549.2.5", "1.2.840.113549.2.7",
                "1.2.840.113549.2.9", "1.2.840.113549.2.10", "1.2.840.113549.2.11", "1.2.840.113549.3.2", "1.2.840.113549.3.7", "1.3.6.1.4.1.311.17.1", "1.3.6.1.4.1.311.17.3.20", "1.3.6.1.4.1.311.20.2.3", "1.3.6.1.4.1.311.88.2.1",
                "1.3.6.1.4.1.311.88.2.2", "1.3.6.1.5.5.7.3.1", "1.3.6.1.5.5.7.3.2", "1.3.6.1.5.5.7.3.3", "1.3.6.1.5.5.7.3.4", "1.3.6.1.5.5.7.3.8", "1.3.6.1.5.5.7.3.9", "1.3.6.1.5.5.7.6.2", "1.3.6.1.5.5.7.48.1", "1.3.6.1.5.5.7.48.1.2",
                "1.3.6.1.5.5.7.48.2", "1.3.14.3.2.26", "1.3.14.3.2.7", "1.3.132.0.34", "1.3.132.0.35", "2.5.4.3", "2.5.4.5", "2.5.4.6", "2.5.4.7", "2.5.4.8", "2.5.4.10", "2.5.4.11", "2.5.4.97", "2.5.29.14", "2.5.29.15", "2.5.29.17", "2.5.29.19",
                "2.5.29.20", "2.5.29.35", "2.16.840.1.101.3.4.1.2", "2.16.840.1.101.3.4.1.22", "2.16.840.1.101.3.4.1.42", "2.16.840.1.101.3.4.2.1", "2.16.840.1.101.3.4.2.2", "2.16.840.1.101.3.4.2.3", "2.23.140.1.2.1", "2.23.140.1.2.2",
            }.Select(s => new KeyValuePair<string, string>(s, s)).ToArray();
        _cd = new ConcurrentDictionary<string, string>(_pairs, IgnoreCase ? StringComparer.Ordinal : StringComparer.OrdinalIgnoreCase);
    }

    [Benchmark]
    public int TryGetValue()
    {
        int count = 0;
        foreach (KeyValuePair<string, string> pair in _pairs)
            if (_cd.TryGetValue(pair.Key, out _)) 
                count++;
        return count;
    }
}

[MemoryDiagnoser(false)]
public class GuidKeys
{
    private KeyValuePair<Guid, int>[] _pairs;
    private ConcurrentDictionary<Guid, int> _cd;

    [GlobalSetup]
    public void Setup()
    {
        _pairs = Enumerable.Range(0, 99).Select(i => new KeyValuePair<Guid, int>(Guid.NewGuid(), i)).ToArray();
        _cd = new ConcurrentDictionary<Guid, int>(_pairs);
    }

    [Benchmark]
    public int TryGetValue()
    {
        int count = 0;
        foreach (KeyValuePair<Guid, int> pair in _pairs)
            if (_cd.TryGetValue(pair.Key, out _)) 
                count++;
        return count;
    }
}

[DisassemblyDiagnoser(maxDepth: 2)]
[MemoryDiagnoser(false)]
public class ObjectKeys
{
    private KeyValuePair<MyObject, int>[] _pairs;
    private ConcurrentDictionary<MyObject, int> _cd;

    [Params(false, true)]
    public bool CustomComparer { get; set; }

    [GlobalSetup]
    public void Setup()
    {
        _pairs = Enumerable.Range(0, 99).Select(i => new KeyValuePair<MyObject, int>(new MyObject(), i)).ToArray();
        _cd = new ConcurrentDictionary<MyObject, int>(_pairs, CustomComparer ? ReferenceEqualityComparer.Instance : EqualityComparer<MyObject>.Default);
    }

    [Benchmark]
    public int TryGetValue()
    {
        int count = 0;
        foreach (KeyValuePair<MyObject, int> pair in _pairs)
            if (_cd.TryGetValue(pair.Key, out _)) 
                count++;
        return count;
    }
}

public class MyObject { }
```