@cincuranet 2023-02-16
Regressions on arm64:
- https://github.com/dotnet/perf-autofiling-issues/issues/13059