@AndyAyersMS 2023-02-16
Possible improvements:
* https://github.com/dotnet/perf-autofiling-issues/issues/13063