@EgorBo 2023-02-21
Improvements on x64: https://github.com/dotnet/perf-autofiling-issues/issues/12986